Create a dedicated HTMLMail object
